Global admins can now edit any weblog
git-svn-id: https://svn.apache.org/repos/asf/incubator/roller/branches/roller_2.0@265546 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/roller/pojos/WebsiteObject.java
@@ -19,7 +19,7 @@ public abstract class WebsiteObject extends PersistentObject
             return true;
         }
         if (getWebsite().hasUserPermissions(
-           roller.getUser(), (short)(PermissionsData.AUTHOR)))
+                roller.getUser(), PermissionsData.AUTHOR))
         {
             return true;
         }
